+++ CHANGELOG.md
@@ -4,3 +4,6 @@
 # Version 1
 
 ## [Unreleased]
+### Ported from [kamilsk/cilex-service-providers](https://github.com/kamilsk/CilexServiceProviders)
+- classes
+  - `OctoLab\Cilex\ServiceProvider\ConfigServiceProvider`